feat: Adds GCD file for updating UI elements
This file and it's function allows UI elements to update
in the Main Queue while keeping network tasks in the background. The
code will be used whenever a UI element should be processed in
the Main Queue while network tasks are running in the background. This
code is used to eliminate concurrency related bugs.